refactor: move Utils into codgen folder
